Changed pipeline file naming lowercase
