+++ tests/mcs_benchmark_data_test/pipelines/CommonsenseQA/commonsense_qa_pipeline_test.py
@@ -5,7 +5,7 @@ from rdflib import Graph
 
 from mcs_benchmark_data.path import DATA_DIR_PATH
 from tests.mcs_benchmark_data_test.assertions import assert_valid_rdf_loaded
-from mcs_benchmark_data.pipelines.CommonsenseQA.commonsense_qa_benchmark_pipeline import (
+from mcs_benchmark_data.pipelines.commonsense_qa.commonsense_qa_benchmark_pipeline import (
     CommonsenseQaBenchmarkPipeline,
 )
 from mcs_benchmark_data.inline_labels_benchmark_file_names import (
