+++ tests/mcs_benchmark_data_test/pipelines/PhysicalIQA/physical_iqa_pipeline_test.py
@@ -1,5 +1,5 @@
 from tests.mcs_benchmark_data_test.assertions import assert_valid_rdf_loaded
-from mcs_benchmark_data.pipelines.PhysicalIQA.physical_iqa_benchmark_pipeline import (
+from mcs_benchmark_data.pipelines.physical_iqa.physical_iqa_benchmark_pipeline import (
     PhysicalIQaBenchmarkPipeline,
 )
 from mcs_benchmark_data.infile_labels_benchmark_file_names import (
