+++ tests/mcs_benchmark_data_test/pipelines/cycic/cycic_pipeline_test.py
@@ -1,5 +1,5 @@
 from tests.mcs_benchmark_data_test.assertions import assert_valid_rdf_loaded
-from mcs_benchmark_data.pipelines.CycIC.cycic_benchmark_pipeline import (
+from mcs_benchmark_data.pipelines.cycic.cycic_benchmark_pipeline import (
     CycicBenchmarkPipeline,
 )
 from mcs_benchmark_data.infile_labels_benchmark_file_names import (
