+++ tests/mcs_benchmark_data_test/pipelines/mcscript/mcscript_pipeline_test.py
@@ -1,5 +1,5 @@
 from tests.mcs_benchmark_data_test.assertions import assert_valid_rdf_loaded
-from mcs_benchmark_data.pipelines.MCScript.mcscript_benchmark_pipeline import (
+from mcs_benchmark_data.pipelines.mcscript.mcscript_benchmark_pipeline import (
     MCScriptBenchmarkPipeline,
 )
 from mcs_benchmark_data.inline_labels_benchmark_file_names import (
